requirements.txt
streamlit
pandas
numpy
requests
st_social_media_links
pillow
ultralytics
plotly
matplotlib
scikit-learn
seaborn
pytorch_lightning
lightning
timm
torchvision
torch
seaborn
beautifulsoup4
niftystocks
ta
mplchart
yfinance
mplfinance
langchain-community
langchain-huggingface
transformers
sentence_transformers
faiss-cpu

